Optimized Huffman table creation resulting in large speed increase.
Instead of creating a lookup table for 15-bit values, the size of the lookup
table now varies with the maximum code length, resulting in much smaller
tables on average, which can be filled much quicker.

Previously, Huffman table creation took around 45% of the decoding time, but
now only a fraction of the total. As a result, the total decode speed increased
by around 40%.